move BYOR functionality to a separate class
