+++ everware/__init__.py
@@ -1,5 +1,6 @@
 __version__ = "0.11.0"
 from .spawner import *
+from .byor_spawner import *
 from .authenticator import *
 from .user_spawn_handler import *
 from .user_wait_handler import *
